@@ -1,14 +1,13 @@
 {
-  "name": "transferkings",
+  "name": "transfer_kings",
   "version": "0.1.0",
   "private": true,
-  "description": "Football Contract Management Game",
+  "description": "JeffBets so you don't have to.",
   "keywords": [
-    "Football",
-    "Soccer",
-    "ICP",
-    "Game",
-    "Play2Earn"
+    "Transfer Kings",
+    "ICFC",
+    "IC",
+    "ICP"
   ],
   "scripts": {
     "dev": "vite dev",
@@ -24,38 +23,38 @@
   "devDependencies": {
     "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
     "@rollup/plugin-inject": "^5.0.5",
-    "@sveltejs/adapter-static": "^3.0.0",
-    "@sveltejs/kit": "^2.0.3",
+    "@sveltejs/adapter-static": "^3.0.8",
+    "@sveltejs/kit": "^2.15.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.4",
     "@types/node": "^20.10.2",
     "@typescript-eslint/eslint-plugin": "^6.9.0",
     "@typescript-eslint/parser": "^6.9.0",
     "autoprefixer": "^10.4.16",
-    "eslint": "^8.16.0",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-svelte": "^2.34.0",
     "sass": "^1.69.5",
-    "svelte": "^4.2.2",
-    "svelte-check": "^3.5.2",
+    "svelte": "^5.15.0",
+    "svelte-check": "^4.0.5",
+    "svelte-preprocess": "^6.0.3",
     "tslib": "^2.6.2",
     "typescript": "^5.2.2",
-    "vite": "^5.3.1"
+    "vite": "^5.4.11"
   },
   "type": "module",
   "dependencies": {
-    "@dfinity/agent": "^1.0.1",
-    "@dfinity/auth-client": "^1.0.1",
-    "@dfinity/candid": "^1.0.1",
-    "@dfinity/gix-components": "^4.1.0",
-    "@dfinity/nns": "^4.0.1",
-    "@dfinity/nns-proto": "^1.0.1",
-    "@dfinity/principal": "^1.0.1",
-    "@dfinity/sns": "^3.0.0",
-    "@dfinity/utils": "^2.1.2",
-    "esbuild": "^0.21.5",
-    "lodash-es": "^4.17.21",
-    "node-modules-polyfill": "^0.1.4",
-    "postcss": "^8.4.31",
-    "svelte-flag-icons": "^0.7.2",
-    "tailwindcss": "^3.3.5"
+    "@dfinity/agent": "^2.1.2",
+    "@dfinity/auth-client": "^2.1.2",
+    "@dfinity/candid": "^2.1.2",
+    "@dfinity/nns": "^7.0.1",
+    "@dfinity/nns-proto": "^2.0.1",
+    "@dfinity/principal": "^2.1.2",
+    "@dfinity/sns": "^3.2.2",
+    "@dfinity/utils": "^2.5.2",
+    "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
+    "dompurify": "^3.1.7",
+    "dotenv": "^16.4.5",
+    "esbuild": "^0.24.0",
+    "postcss": "^8.4.47",
+    "svelte-flag-icons": "^1.0.3",
+    "tailwindcss": "^3.4.14"
   }
 }
