@@ -1,33 +1,47 @@
 
-import root from '../root.svelte';
+import root from '../root.js';
 import { set_building, set_prerendering } from '__sveltekit/environment';
 import { set_assets } from '__sveltekit/paths';
 import { set_manifest, set_read_implementation } from '__sveltekit/server';
 import { set_private_env, set_public_env, set_safe_public_env } from '../../../node_modules/@sveltejs/kit/src/runtime/shared-server.js';
 
 export const options = {
-	app_dir: "_app",
 	app_template_contains_nonce: false,
 	csp: {"mode":"auto","directives":{"upgrade-insecure-requests":false,"block-all-mixed-content":false},"reportOnly":{"upgrade-insecure-requests":false,"block-all-mixed-content":false}},
 	csrf_check_origin: true,
 	embedded: false,
 	env_public_prefix: 'PUBLIC_',
 	env_private_prefix: '',
+	hash_routing: false,
 	hooks: null, // added lazily, via `get_hooks`
 	preload_strategy: "modulepreload",
 	root,
 	service_worker: false,
 	templates: {
-		app: ({ head, body, assets, nonce, env }) => "<!doctype html>\n<html lang=\"en\">\n  <head>\n    <meta charset=\"utf-8\" />\n    <meta content=\"width=device-width, initial-scale=1\" name=\"viewport\" />\n\n    <title>Transfer Kings</title>\n    <link href=\"https://transferkings.xyz\" rel=\"canonical\" />\n    <meta\n      content=\"Play Transfer Kings on the Internet Computer blockchain and become football agent today.\"\n      name=\"description\"\n    />\n    <meta content=\"Transfer Kings\" property=\"og:title\" />\n    <meta\n      content=\"Play Transfer Kings on the Internet Computer blockchain and become football agent today.\"\n      property=\"og:description\"\n    />\n    <meta content=\"website\" property=\"og:type\" />\n    <meta content=\"https://transferkings.xyz\" property=\"og:url\" />\n    <meta content=\"https://transferkings.xyz/meta-share.jpg\" property=\"og:image\" />\n    <meta content=\"summary_large_image\" name=\"twitter:card\" />\n    <meta content=\"Transfer Kings\" name=\"twitter:title\" />\n    <meta\n      content=\"Play Transfer Kings on the Internet Computer blockchain and become football agent today.\"\n      name=\"twitter:description\"\n    />\n    <meta\n      content=\"https://transferkings.xyz/meta-share.jpg\"\n      name=\"twitter:image\"\n    />\n    <meta content=\"@beadle1989\" name=\"twitter:creator\" />\n\n    <link crossorigin=\"anonymous\" href=\"/manifest.webmanifest\" rel=\"manifest\" />\n\n    <!-- Favicon -->\n    <link\n      rel=\"icon\"\n      type=\"image/png\"\n      sizes=\"32x32\"\n      href=\"" + assets + "/favicons/favicon-32x32.png\"\n    />\n    <link\n      rel=\"icon\"\n      type=\"image/png\"\n      sizes=\"16x16\"\n      href=\"" + assets + "/favicons/favicon-16x16.png\"\n    />\n    <link rel=\"shortcut icon\" href=\"" + assets + "/favicons/favicon.ico\" />\n\n    <!-- iOS meta tags & icons -->\n    <meta name=\"apple-mobile-web-app-capable\" content=\"yes\" />\n    <meta name=\"apple-mobile-web-app-status-bar-style\" content=\"#2CE3A6\" />\n    <meta name=\"apple-mobile-web-app-title\" content=\"Transfer Kings\" />\n    <link\n      rel=\"apple-touch-icon\"\n      href=\"" + assets + "/favicons/apple-touch-icon.png\"\n    />\n    <link\n      rel=\"mask-icon\"\n      href=\"" + assets + "/favicons/safari-pinned-tab.svg\"\n      color=\"#2CE3A6\"\n    />\n\n    <!-- MS -->\n    <meta name=\"msapplication-TileColor\" content=\"#101111\" />\n    <meta\n      name=\"msapplication-config\"\n      content=\"" + assets + "/favicons/browserconfig.xml\"\n    />\n\n    <meta content=\"#2CE3A6\" name=\"theme-color\" />\n    " + head + "\n\n    <style>\n      html,\n      body {\n        height: 100%;\n        margin: 0;\n      }\n\n      @font-face {\n        font-display: swap;\n        font-family: \"Rubik\";\n        font-style: normal;\n        font-weight: 400;\n        src: url(\"" + assets + "/Rubik-Regular.woff2\") format(\"woff2\");\n      }\n      body {\n        font-family: \"Rubik\", sans-serif !important;\n        color: white !important;\n        height: 100vh;\n        margin: 0;\n        background-image: url(\"/background.jpg\");\n        background-size: cover;\n        background-position: center;\n        background-repeat: no-repeat;\n        background-attachment: fixed;\n      }\n\n      #app-spinner {\n        --spinner-size: 30px;\n\n        width: var(--spinner-size);\n        height: var(--spinner-size);\n\n        animation: app-spinner-linear-rotate 2000ms linear infinite;\n\n        position: absolute;\n        top: calc(50% - (var(--spinner-size) / 2));\n        left: calc(50% - (var(--spinner-size) / 2));\n\n        --radius: 45px;\n        --circumference: calc(3.14159265359 * var(--radius) * 2);\n\n        --start: calc((1 - 0.05) * var(--circumference));\n        --end: calc((1 - 0.8) * var(--circumference));\n      }\n\n      #app-spinner circle {\n        stroke-dasharray: var(--circumference);\n        stroke-width: 10%;\n        transform-origin: 50% 50% 0;\n\n        transition-property: stroke;\n\n        animation-name: app-spinner-stroke-rotate-100;\n        animation-duration: 4000ms;\n        animation-timing-function: cubic-bezier(0.35, 0, 0.25, 1);\n        animation-iteration-count: infinite;\n\n        fill: transparent;\n        stroke: currentColor;\n\n        transition: stroke-dashoffset 225ms linear;\n      }\n\n      @keyframes app-spinner-linear-rotate {\n        0% {\n          transform: rotate(0deg);\n        }\n        100% {\n          transform: rotate(360deg);\n        }\n      }\n\n      @keyframes app-spinner-stroke-rotate-100 {\n        0% {\n          stroke-dashoffset: var(--start);\n          transform: rotate(0);\n        }\n        12.5% {\n          stroke-dashoffset: var(--end);\n          transform: rotate(0);\n        }\n        12.5001% {\n          stroke-dashoffset: var(--end);\n          transform: rotateX(180deg) rotate(72.5deg);\n        }\n        25% {\n          stroke-dashoffset: var(--start);\n          transform: rotateX(180deg) rotate(72.5deg);\n        }\n\n        25.0001% {\n          stroke-dashoffset: var(--start);\n          transform: rotate(270deg);\n        }\n        37.5% {\n          stroke-dashoffset: var(--end);\n          transform: rotate(270deg);\n        }\n        37.5001% {\n          stroke-dashoffset: var(--end);\n          transform: rotateX(180deg) rotate(161.5deg);\n        }\n        50% {\n          stroke-dashoffset: var(--start);\n          transform: rotateX(180deg) rotate(161.5deg);\n        }\n\n        50.0001% {\n          stroke-dashoffset: var(--start);\n          transform: rotate(180deg);\n        }\n        62.5% {\n          stroke-dashoffset: var(--end);\n          transform: rotate(180deg);\n        }\n        62.5001% {\n          stroke-dashoffset: var(--end);\n          transform: rotateX(180deg) rotate(251.5deg);\n        }\n        75% {\n          stroke-dashoffset: var(--start);\n          transform: rotateX(180deg) rotate(251.5deg);\n        }\n\n        75.0001% {\n          stroke-dashoffset: var(--start);\n          transform: rotate(90deg);\n        }\n        87.5% {\n          stroke-dashoffset: var(--end);\n          transform: rotate(90deg);\n        }\n        87.5001% {\n          stroke-dashoffset: var(--end);\n          transform: rotateX(180deg) rotate(341.5deg);\n        }\n        100% {\n          stroke-dashoffset: var(--start);\n          transform: rotateX(180deg) rotate(341.5deg);\n        }\n      }\n    </style>\n  </head>\n  <body data-sveltekit-preload-data=\"hover\">\n    <div style=\"display: contents\">" + body + "</div>\n\n    <svg\n      id=\"app-spinner\"\n      preserveAspectRatio=\"xMidYMid meet\"\n      focusable=\"false\"\n      aria-hidden=\"true\"\n      data-tid=\"spinner\"\n      viewBox=\"0 0 100 100\"\n    >\n      <circle cx=\"50%\" cy=\"50%\" r=\"45\" />\n    </svg>\n  </body>\n</html>\n",
+		app: ({ head, body, assets, nonce, env }) => "<!doctype html>\n<html lang=\"en\">\n  <head>\n    <meta charset=\"utf-8\" />\n    <meta content=\"width=device-width, initial-scale=1\" name=\"viewport\" />\n\n    <title>Transfer Kings</title>\n    <link href=\"https://transferkings.xyz\" rel=\"canonical\" />\n    <meta\n      content=\"Collect your favourite players on Transfer Kings.\"\n      name=\"description\"\n    />\n    <meta\n      content=\"Collect your favourite players on Transfer Kings\"\n      property=\"og:title\"\n    />\n    <meta\n      content=\"Collect your favourite players on Transfer Kings.\"\n      property=\"og:description\"\n    />\n    <meta content=\"website\" property=\"og:type\" />\n    <meta content=\"https://transferkings.xyz\" property=\"og:url\" />\n    <meta\n      content=\"https://transferkings.xyz/meta-share.jpg\"\n      property=\"og:image\"\n    />\n    <meta content=\"summary_large_image\" name=\"twitter:card\" />\n    <meta content=\"Transfer Kings\" name=\"twitter:title\" />\n    <meta\n      content=\"Collect your favourite players on Transfer Kings.\"\n      name=\"twitter:description\"\n    />\n    <meta\n      content=\"https://transferkings.xyz/meta-share.jpg\"\n      name=\"twitter:image\"\n    />\n    <meta content=\"@beadle1989\" name=\"twitter:creator\" />\n\n    <link crossorigin=\"anonymous\" href=\"/manifest.webmanifest\" rel=\"manifest\" />\n\n    <link\n      rel=\"icon\"\n      type=\"image/png\"\n      sizes=\"32x32\"\n      href=\"" + assets + "/favicons/favicon-32x32.png\"\n    />\n    <link\n      rel=\"icon\"\n      type=\"image/png\"\n      sizes=\"16x16\"\n      href=\"" + assets + "/favicons/favicon-16x16.png\"\n    />\n    <link rel=\"shortcut icon\" href=\"" + assets + "/favicons/favicon.ico\" />\n\n    <meta name=\"mobile-web-app-capable\" content=\"yes\" />\n    <meta name=\"apple-mobile-web-app-status-bar-style\" content=\"#D7FE49\" />\n    <meta name=\"apple-mobile-web-app-title\" content=\"Transfer Kings\" />\n    <link\n      rel=\"apple-touch-icon\"\n      href=\"" + assets + "/favicons/apple-touch-icon.png\"\n    />\n    <link\n      rel=\"mask-icon\"\n      href=\"" + assets + "/favicons/safari-pinned-tab.svg\"\n      color=\"#D7FE49\"\n    />\n\n    <meta name=\"msapplication-TileColor\" content=\"#D7FE49\" />\n    <meta\n      name=\"msapplication-config\"\n      content=\"" + assets + "/favicons/browserconfig.xml\"\n    />\n\n    <meta content=\"#D7FE49\" name=\"theme-color\" />\n    " + head + "\n\n    <style>\n      html,\n      body {\n        height: 100%;\n        margin: 0;\n      }\n\n      @font-face {\n        font-display: swap;\n        font-family: \"Tanker\";\n        font-style: normal;\n        font-weight: 400;\n        src: url(\"" + assets + "/Tanker-Regular.woff2\") format(\"woff2\");\n      }\n      body {\n        font-family: \"Tanker\", sans-serif !important;\n        color: white !important;\n        background-color: #0d0d0d;\n        height: 100vh;\n        margin: 0;\n      }\n\n      #app-spinner {\n        --spinner-size: 30px;\n\n        width: var(--spinner-size);\n        height: var(--spinner-size);\n\n        animation: app-spinner-linear-rotate 2000ms linear infinite;\n\n        position: absolute;\n        top: calc(50% - (var(--spinner-size) / 2));\n        left: calc(50% - (var(--spinner-size) / 2));\n\n        --radius: 45px;\n        --circumference: calc(3.14159265359 * var(--radius) * 2);\n\n        --start: calc((1 - 0.05) * var(--circumference));\n        --end: calc((1 - 0.8) * var(--circumference));\n      }\n\n      #app-spinner circle {\n        stroke-dasharray: var(--circumference);\n        stroke-width: 10%;\n        transform-origin: 50% 50% 0;\n\n        transition-property: stroke;\n\n        animation-name: app-spinner-stroke-rotate-100;\n        animation-duration: 4000ms;\n        animation-timing-function: cubic-bezier(0.35, 0, 0.25, 1);\n        animation-iteration-count: infinite;\n\n        fill: transparent;\n        stroke: currentColor;\n\n        transition: stroke-dashoffset 225ms linear;\n      }\n\n      @keyframes app-spinner-linear-rotate {\n        0% {\n          transform: rotate(0deg);\n        }\n        100% {\n          transform: rotate(360deg);\n        }\n      }\n\n      @keyframes app-spinner-stroke-rotate-100 {\n        0% {\n          stroke-dashoffset: var(--start);\n          transform: rotate(0);\n        }\n        12.5% {\n          stroke-dashoffset: var(--end);\n          transform: rotate(0);\n        }\n        12.5001% {\n          stroke-dashoffset: var(--end);\n          transform: rotateX(180deg) rotate(72.5deg);\n        }\n        25% {\n          stroke-dashoffset: var(--start);\n          transform: rotateX(180deg) rotate(72.5deg);\n        }\n\n        25.0001% {\n          stroke-dashoffset: var(--start);\n          transform: rotate(270deg);\n        }\n        37.5% {\n          stroke-dashoffset: var(--end);\n          transform: rotate(270deg);\n        }\n        37.5001% {\n          stroke-dashoffset: var(--end);\n          transform: rotateX(180deg) rotate(161.5deg);\n        }\n        50% {\n          stroke-dashoffset: var(--start);\n          transform: rotateX(180deg) rotate(161.5deg);\n        }\n\n        50.0001% {\n          stroke-dashoffset: var(--start);\n          transform: rotate(180deg);\n        }\n        62.5% {\n          stroke-dashoffset: var(--end);\n          transform: rotate(180deg);\n        }\n        62.5001% {\n          stroke-dashoffset: var(--end);\n          transform: rotateX(180deg) rotate(251.5deg);\n        }\n        75% {\n          stroke-dashoffset: var(--start);\n          transform: rotateX(180deg) rotate(251.5deg);\n        }\n\n        75.0001% {\n          stroke-dashoffset: var(--start);\n          transform: rotate(90deg);\n        }\n        87.5% {\n          stroke-dashoffset: var(--end);\n          transform: rotate(90deg);\n        }\n        87.5001% {\n          stroke-dashoffset: var(--end);\n          transform: rotateX(180deg) rotate(341.5deg);\n        }\n        100% {\n          stroke-dashoffset: var(--start);\n          transform: rotateX(180deg) rotate(341.5deg);\n        }\n      }\n    </style>\n  </head>\n  <body data-sveltekit-preload-data=\"hover\">\n    <div style=\"display: contents\">" + body + "</div>\n\n    <svg\n      id=\"app-spinner\"\n      preserveAspectRatio=\"xMidYMid meet\"\n      focusable=\"false\"\n      aria-hidden=\"true\"\n      data-tid=\"spinner\"\n      viewBox=\"0 0 100 100\"\n    >\n      <circle cx=\"50%\" cy=\"50%\" r=\"45\" />\n    </svg>\n  </body>\n</html>\n",
 		error: ({ status, message }) => "<!doctype html>\n<html lang=\"en\">\n\t<head>\n\t\t<meta charset=\"utf-8\" />\n\t\t<title>" + message + "</title>\n\n\t\t<style>\n\t\t\tbody {\n\t\t\t\t--bg: white;\n\t\t\t\t--fg: #222;\n\t\t\t\t--divider: #ccc;\n\t\t\t\tbackground: var(--bg);\n\t\t\t\tcolor: var(--fg);\n\t\t\t\tfont-family:\n\t\t\t\t\tsystem-ui,\n\t\t\t\t\t-apple-system,\n\t\t\t\t\tBlinkMacSystemFont,\n\t\t\t\t\t'Segoe UI',\n\t\t\t\t\tRoboto,\n\t\t\t\t\tOxygen,\n\t\t\t\t\tUbuntu,\n\t\t\t\t\tCantarell,\n\t\t\t\t\t'Open Sans',\n\t\t\t\t\t'Helvetica Neue',\n\t\t\t\t\tsans-serif;\n\t\t\t\tdisplay: flex;\n\t\t\t\talign-items: center;\n\t\t\t\tjustify-content: center;\n\t\t\t\theight: 100vh;\n\t\t\t\tmargin: 0;\n\t\t\t}\n\n\t\t\t.error {\n\t\t\t\tdisplay: flex;\n\t\t\t\talign-items: center;\n\t\t\t\tmax-width: 32rem;\n\t\t\t\tmargin: 0 1rem;\n\t\t\t}\n\n\t\t\t.status {\n\t\t\t\tfont-weight: 200;\n\t\t\t\tfont-size: 3rem;\n\t\t\t\tline-height: 1;\n\t\t\t\tposition: relative;\n\t\t\t\ttop: -0.05rem;\n\t\t\t}\n\n\t\t\t.message {\n\t\t\t\tborder-left: 1px solid var(--divider);\n\t\t\t\tpadding: 0 0 0 1rem;\n\t\t\t\tmargin: 0 0 0 1rem;\n\t\t\t\tmin-height: 2.5rem;\n\t\t\t\tdisplay: flex;\n\t\t\t\talign-items: center;\n\t\t\t}\n\n\t\t\t.message h1 {\n\t\t\t\tfont-weight: 400;\n\t\t\t\tfont-size: 1em;\n\t\t\t\tmargin: 0;\n\t\t\t}\n\n\t\t\t@media (prefers-color-scheme: dark) {\n\t\t\t\tbody {\n\t\t\t\t\t--bg: #222;\n\t\t\t\t\t--fg: #ddd;\n\t\t\t\t\t--divider: #666;\n\t\t\t\t}\n\t\t\t}\n\t\t</style>\n\t</head>\n\t<body>\n\t\t<div class=\"error\">\n\t\t\t<span class=\"status\">" + status + "</span>\n\t\t\t<div class=\"message\">\n\t\t\t\t<h1>" + message + "</h1>\n\t\t\t</div>\n\t\t</div>\n\t</body>\n</html>\n"
 	},
-	version_hash: "fv45wz"
+	version_hash: "vfkn6m"
 };
 
 export async function get_hooks() {
+	let handle;
+	let handleFetch;
+	let handleError;
+	let init;
+	
+
+	let reroute;
+	let transport;
+	
+
 	return {
-		
-		
+		handle,
+		handleFetch,
+		handleError,
+		init,
+		reroute,
+		transport
 	};
 }
 
