@@ -1,6 +1,5 @@
-import { P } from "../../chunks/index.js";
-import "@dfinity/auth-client";
-import "@dfinity/utils";
+import "clsx";
+import { _ } from "../../chunks/index.js";
 export {
-  P as default
+  _ as default
 };
