@@ -2,7 +2,7 @@
 
 export const index = 3;
 let component_cache;
-export const component = async () => component_cache ??= (await import('../entries/pages/agent-hub/_page.svelte.js')).default;
-export const imports = ["_app/immutable/nodes/3.Ch3-azL_.js","_app/immutable/chunks/index.BZ8X_3bo.js","_app/immutable/chunks/vendor.ThHMLpiP.js"];
-export const stylesheets = ["_app/immutable/assets/index.C-Mw7l9i.css"];
+export const component = async () => component_cache ??= (await import('../entries/pages/profile/_page.svelte.js')).default;
+export const imports = ["_app/immutable/nodes/3.CANNjDQr.js","_app/immutable/chunks/B47vXuiC.js","_app/immutable/chunks/DJH_4bDg.js"];
+export const stylesheets = ["_app/immutable/assets/index.DtggSJ5a.css"];
 export const fonts = [];
