@@ -1,22 +1,6 @@
-import { authStore } from "$lib/stores/auth-store";
+import { authStore } from "../stores/auth-store";
 import { derived, type Readable } from "svelte/store";
-
-/*LOCALDEVONLY
-const adminPrincipal =
-  "opyzn-r7zln-jwgvb-tx75c-ncekh-xhvje-epcj7-saonq-z732m-zi4mm-qae";
-*/
-
-const adminPrincipal =
-  "nn75s-ayupf-j6mj3-kluyb-wjj7y-eang2-dwzzr-cfdxk-etbw7-cgwnb-lqe";
 export const authSignedInStore: Readable<boolean> = derived(
   authStore,
   ({ identity }) => identity !== null && identity !== undefined,
-);
-
-export const authIsAdmin: Readable<boolean> = derived(
-  authStore,
-  ({ identity }) =>
-    identity !== null &&
-    identity !== undefined &&
-    identity.getPrincipal().toString() === adminPrincipal,
 );
