@@ -6,6 +6,7 @@ import { dirname, join, resolve } from "path";
 import { fileURLToPath } from "url";
 import type { UserConfig } from "vite";
 import { defineConfig, loadEnv } from "vite";
+import type { PluginBuild } from "esbuild";
 
 const file = fileURLToPath(new URL("package.json", import.meta.url));
 const json = readFileSync(file, "utf8");
@@ -61,8 +62,6 @@ const config: UserConfig = {
     preprocessorOptions: {
       scss: {
         additionalData: `
-          @use "./node_modules/@dfinity/gix-components/dist/styles/mixins/media";
-          @use "./node_modules/@dfinity/gix-components/dist/styles/mixins/text";
         `,
       },
     },
@@ -77,8 +76,8 @@ const config: UserConfig = {
           const lazy = ["@dfinity/nns"];
 
           if (
-            ["@sveltejs", "svelte", "@dfinity/gix-components", ...lazy].find(
-              (lib) => folder.includes(lib),
+            ["@sveltejs", "svelte", ...lazy].find((lib) =>
+              folder.includes(lib),
             ) === undefined &&
             folder.includes("node_modules")
           ) {
@@ -100,6 +99,15 @@ const config: UserConfig = {
         inject({
           modules: { Buffer: ["buffer", "Buffer"] },
         }),
+        {
+          name: "fix-node-globals-polyfill",
+          setup(build: PluginBuild) {
+            build.onResolve(
+              { filter: /_virtual-process-polyfill_\.js/ },
+              ({ path }) => ({ path }),
+            );
+          },
+        } as any, // Type assertion to bypass the type mismatch
       ],
     },
   },
@@ -123,13 +131,13 @@ const config: UserConfig = {
         NodeModulesPolyfillPlugin(),
         {
           name: "fix-node-globals-polyfill",
-          setup(build) {
+          setup(build: PluginBuild) {
             build.onResolve(
               { filter: /_virtual-process-polyfill_\.js/ },
               ({ path }) => ({ path }),
             );
           },
-        },
+        } as any, // Type assertion to bypass the type mismatch
       ],
     },
   },
@@ -160,9 +168,13 @@ export default defineConfig((): UserConfig => {
       "process.env": {
         ...readCanisterIds({}),
         DFX_NETWORK: network,
+        DATA_CANISTER_ID: "52fzd-2aaaa-aaaal-qmzsa-cai",
       },
       VITE_APP_VERSION: JSON.stringify(version),
       VITE_DFX_NETWORK: JSON.stringify(network),
+      "process.env.CANISTER_ID_BACKEND": JSON.stringify(
+        process.env.CANISTER_ID_BACKEND,
+      ),
     },
   };
 });
